requirements.txt
SpeechRecognition
pydub
openai
python-dotenv
gTTS
python-vlc
PyAudio